Refactor : `Writer` 생성
- article과 Member 가지고 있는 Writer 추가.
