+++ src/main/java/kr/codesqaud/cafe/controller/MemberController.java
@@ -6,14 +6,12 @@ import kr.codesqaud.cafe.exception.*;
 import kr.codesqaud.cafe.repository.MemberRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
 import java.util.List;
-import java.util.Optional;
 
 import static kr.codesqaud.cafe.constant.ConstUrl.REDIRECT_INDEX;
 import static kr.codesqaud.cafe.exception.InvalidAuthorityException.INVALID_MEMBER;
